packetio: handle not-int REST id's properly
We expect REST id's to be integers.  If they are not, then attempting
to fetch them as an integer type from the Pistache request object will
throw an exception.  This change adds code to capture the
exception and return a proper HTTP response in its place.

Also do some cleanup while here.

Note that Pistache wraps the handler functions in a function of it's own
that always returns the Route::Result::Ok value, so our handlers don't
need to (and in fact, anything they return will just be ignored).