framework: fix module registration ordering bug
The ordering of static object initialization is at the discretion of the
linker.  With some linkers, the existing module registration framework
could cause objects to be dropped from the modules list due to
initialization of the list object happening *after* the object was
inserted into the list.  In that case, the next ptr would be initialized
(back) to null which would drop subsequent items off the list.

To fix this, update the registration macro to explicitly initialize the
module struct before it can be inserted into the list.

Update the options structure to no longer accept a 'void*' parameter.
No client used it and it could cause a similar problem in the future.
Plain 'ol data is less susceptible to these types of init problems.

Additionally, simplify both modules and options to use the simplest list
type available.  Since we're only iterating, fancy list implementations
don't matter.

Finally, add assertions to try to catch this issue in the future.