framework: add icp_halt and supporting functions
The icp_halt() function provides a way for clients to cleanly
shutdown all modules as well as the ZeroMQ context for the
program.

This change adds a new regiration callback function for modules:
'finish'.  Modules can use this to stop their services and
prepare for shutdown.

Additionally, fix some minor issues that prevented clean shutdown:
* Do better error checking on some zmq calls
* Add a virtual destructor to the api handlers base class so that
  derived destructors can be called
* Work around the Pistache server hanging it's thread on shutdown.
* Listen for SIGTERM (in addition to SIGINT) when blocked on epoll.

With these changes, a single SIGINT or SIGTERM is sufficient to cleanly
shutdown the binary.