tests/aat: update stack command line
Due to the fact that lwIP pbufs are embedded into DPDK mbufs, there
is a race between tx and rx pbufs when the same mbuf is both
transmitted and received by the same stack. This results in the
same pbuf being used simultaneously by both rx and tx functions, which
can cause an assertion failure in the stack code due to invalid
reference counts.

Note, this is only an issue because the AATs use DPDK rings as
interfaces, so the same mbuf (and hence pbuf) appears on both the tx and
rx sides.  Normally, this wouldn't happen.

As a result, update the command line to contain a core mask that will
prevent the stack from actually running worker threads.  Without
workers, the packets on the ring are never received, and the data race
is prevented.