modules/socket: refactor dgram channels
This change simplifies the use of the dgram channel by reducing
the implementation to a pair of bipartite rings.  By bipartite, we mean each
item in the ring either belongs to the stack or the client, with a clear
demarcation between who owns what.  In this case, each item is a packet
buffer capable of containing a full packet + addressing info.

This design is influenced by the way netmap client access netmap ring
data.

Add a unit test to verify bipartite_ring behavior.