modules/packetio: simplify tx queue
In an attempt to send multiple packets per burst, the tx queue
implementation used timerfd's to wake up workers after a short delay
when the stack enqueued packets for transmit.

With the addition of GSO, this delay is no longer necessary.  So replace
the timerfd with an eventfd.  This makes the notification process much
simpler and removes latency from the tx path.