modules/packetio: fix an intermittent assertion failure
The lwIP TCP PCB keeps track of an "oversize" value that stores the
amount of extra room available in a segment.  Various assertions
insist that this value must be less than one MSS.  Since our packet
buffers are larger than this, it's possible for them to have more
room available that lwIP thinks is possible.  To work around this,
clamp the oversize value to no more than one MSS.