modules/socket: improve TCP transmit function
The lwip tcp_write function takes a uint16_t as input, however the
underlying send buffer may actually be larger.  This change modifies the
TCP transmit function so that it checks the *actual* size of the send
buffer and then breaks the actual tcp_write calls into uint16_t chunks.

Additionally, this change modifes the tcp_output calling behavior.  Now
the transmit function only calls tcp_output when the tx buffer is cleared
or the stack's send buffer is full.