modules/packetio: use DPDK memcpy if available
The lwip stack allows platforms to override the default libc memcpy if
desired, so take advantage of it when using the DPDK backend.  There's a
slight benefit to this as the DPDK memcpy uses SSE4 whereas libc currently
only uses SSE3.

Also, increase the pbuf header size to a full cache line to increase
the likelihoood that the packet payload is nicely aligned.