added ICP_BINDTODEVICE=ioN to bind socket to interface in envvar
AF_INET sockets created in ICP shim-layer enabled binaries will be bound (SO_BINDTODEVICE) to
the interface specified in ICP_BINDTODEVICE when this variable exists in the environment. If
the interface does not exist, the socket is closed and -1 is returned with errno = ENODEV.
This feature was added to support communication between multiple ICP shim-layer enabled binaries.