Inception cleanup changes (#38)
1. Add fatal error message if API server socket or shared memory segment already
exist on startup.
2. Add `--force-unlink` command line option to remove files from item 1 above.
3. Shift cleanup from AAT framework to AAT config file using `--force-unlink` option.

Item 3 is a temporary work around for an intermittent issue where Inception does not
shut down cleanly after running the AATs.

To safely implement item 2 above, this change also adds support for long-form arguments that do not have a corresponding short-form one. In the case of `--force-unlink` there is no single character version.