modules/socket: optimize the tcp socket pbuf queue
This change improves the cache utilization of the pbuf queue.  When
pbufs chains, in the form of tcp segments, are added to the queue, all
of the pbufs in the chain are likely to be in the CPU cache.  When the
client decides to read the pbuf payload data at a later time, those same
pbufs are not likely to be in the cache anymore.  So, instead of walking
the pbuf chain linked list when the client wants data, walk the list
when the pbufs are added to the queue (and hence still in the CPU
cache).  Additionally, break the chains up so we no longer have to walk
the linked list in the future. This has the added benefit of
significantly simplifying the underling logic used in generating iovecs
for writing to the buffer and clearing the queue.

Finally, add a max length parameter to the queue's iovec function so
that we can stop as soon as we hit the caller's send limit.