modules/socket: add optimized write for iovecs
The tcpip thread is effectively memory I/O bound on the receive path.
This change adds an optimized write for iovecs to the
circular_buffer_producer template to lessen the overhead of writing from
the stack to the socket's shared memory buffer.

Additionally, always report the writable size of the buffer as a
multiple of 32 bytes.  This reduces the number of unaligned writes to
the buffer when using our AVX memcpy routines.

Finally, update the stream channel to use the optimized write function.