modules/socket: set the PSH flag less often
The TCP PSH flag is used to indicate that the "receiver should pass data
up the stack without delay".  We use it to initiate a copy from the
stack to the stream channel.  Thus, setting this flag unnecessarily
causes our receive code to initiate an unnecessary, and hence
inefficient, write to the user.

This change lessens the number of times we set the PSH flag on outgoing
packets by allowing the tcp transmit caller to specify a function to
check whether the flag needs to be set or not when the final write of
the tranmit is performed.

Additionally, this change simplifies some code logic and lowers some
loop processing thresholds to increase the interleaving between stack
activities and data copying.