modules/socket: fix gso utils bug with partial ACKs
In certain cases, the pbuf_drop_at function adjusted the payload pointer
of the first pbuf in the chain, which pointed to the packet headers.
Higher level callers, with their own pointers to the headers, were
unaware of this update which caused confusion and delay in the callers.

Update the function to move the data and not the payload pointer.