modules/packetio: refactor dpdk worker interrupts
Add a 0MQ socket wrapper that allows the 0MQ control socket to be added
to DPDK's epoll loop.  This allows us to stop polling the control socket
periodically and only wake up when there are actual events to service.

Additionally, stop idle spinning after servicing the queues.  For our
current uses cases, the spinning is rarely if ever beneficial.