modules/socket: allow clients to map shared memory anywhere
Originally, the client was required to mmap the server's shared memory
segment into the same address range as the server because the client and
server shared pointers to channel objects in the shared memory space.

This changelist makes two major changes that allow clients to map the
shared memory segment to any address range and eliminates intermittent
client errors caused by the inability to do so:

1) Pass the channel's shared memory offset, instead of a pointer, when
   setting up sockets.
2) Replace the stream channel buffer pointers with offset pointers.

Additionally, this change also removes the shared segment code related
to mapping segments to specific addresses.