modules/packetio: copy, don't clone, mbufs in EAL tx copy function
The eal_tx_burst_copy_function is intended to allow the stack to
transmit and receive data over DPDK ring interfaces.  However, an mbuf
clone is insufficient for this purpose.  Whle cloned mbufs have
independent mbuf headers, and hence independent pbuf data, clones do not
have independent payload data.  Thus, when transmitted over a ring, both
the source and destination sockets have pointers to the same payload
data inside the original mbuf.  This is especially problematic for TCP
sockets, as the sender must maintain, reference, and possibly modify
transmitted data in order to process incoming ACKs.

Hence, the only way for the stack to operate in a back-to-back manner
using DPDK ring devices is for both the sender and the receiver to have
independent copies of packet data.