packetio/stack: fixes bug/issue #39 causing unclean shutdown
The "tcpip_thread" requires an explicit shutdown notification
so its resources can be closed/cleaned-up. No such notification
existed and cleanup depended on the thread being in a certain
event loop state when the termination signal was caught.

This gap is closed by notifying the thread it needs to shutdown
and exit its forever loop causing the destructor to be called.

A simple "tcpip_shutdown()" function was added and called by lwip
destructor. The destructor then waits for termination of tcpip_thread.

Closes #39