modules/packetio: add PGA API framework
This change defines two new generic objects: sinks and sources.

A generic sink consumes packets for analysis or capture.
A generic sink produces packets for transmission.

Additionally, this change adds a framework to allow a client to attach
and detach generic sinks and sources to packetio internals.  These
functions are currently stubbed out and will be implemented in a
subsequent commit.