modules/packetio: remove PGA; use internal instead
Remove the PGA client/server code in favor of extending the internal
client/server functionality.

Update the internal client/server to handle generic sink and source
related requests.  Note that while the client functions exist, the
server functions are still stubbed out.  We need to add some additional
infrastructure in order to use sources and sinks in the workers.