modules/packetio: add a RCU mechanism
This change adds a new object, recycle::depot(), which contains the
machinery needed to perform Quiescent State Based Reclamation (QSBR)
based deletion of objects shared between a writer and an arbitrary
number of readers.

Use this new object in the workers to provide a way to cleanly remove
data that is shared between them.