modules/packetio: replace vif_map with forwarding_table
The forwarding table extends the functionality of the vif_map by adding
the ability to store both sources and sinks in addition to interface
pointers.  The forwarding table interface is also compatible with the
recently added QSBR memory reclamation mechanism to allow removed
entries to be deleted when no longer needed.  Finally, the internal data
structures are built using the recently added immutable data structures
from the immer package.