modules/packetio: simplify forwarding table key
Since the forwarding table now stores per port interface maps, we don't
need to bother with storing the port in addition to the mac as the key.
This was a unnecessary holdover from the original vif_map
implementation.

Note: also add hash support for the mac_address type, as it is required
when used as a map key.