modules/packetio: add memory pools to generic sources
In order to actually send packets, a generic_source needs a source for
raw packets to modify. To accomplish that, create a backend specific
memory pool for each generic packet source when it is added to the
back end.  Additionally, create a new structure, tx_source, that binds
the memory pool to the generic source.  This allows the scheduler to
seamlessly generate packets as needed.

With this change, generic packet sources can actually transmit packets
on the wire.