modules/packetio: fix DPDK default pool allocation
The pool allocation constructor neglected to consider the number
of transmit and receive queues when calculating the number of packets
needed for the pool.  Update the constructor to accept another parameter
which contains the port queue counts so that it can do so.

Note that we cannot just use the port info vector to retrieve the queue
counts as the queues have not been configured when the buffers need to
be allocated...  There is a chicken/egg problem here.

Update the queue utils compilation unit to include a common queue
counting function and remove redunant copies.