build: add proper dependencies for versioned files
Multiple files within this project use build variables as dependencies
to hard code version information into the objects at compile time.
Previously, those files were compiled at every invocation of `make`.
This change updates the build infrastructure so that the files that
contain the version information will only be built if there are actual
changes to the sources that the version information represents.