modules/socket: fix circular buffer wrapping bug
The circular buffer producer template, as used by the stream channel,
lacked a check to prevent writes that spanned the end of the buffer from
exceeding the buffer's available capacity.  This resulted in data
disappearing from the buffer in certain cases.  This fix adds a check to
make sure that there is sufficient space in the buffer before writing data
across the wrap.

Additionally, this change affects push notification logic in the tcp
socket, so update the `buffer empty` function to return true if the TCP
socket writes all of the peekable data instead of all of the readable
data.