tests: warn about untested spirent_pga functions
If the host machine running the unit tests cannot test some of the
vector functions due to a lack of host CPU support, generate a warning.

Also, use functions instead of static lists for instruction sets.  This
makes the code a little neater and adds the ability to easily test other
host CPU architectures, if needed.