mk: allow command line compiler optimization flags
This change allows users to specify explicit compiler optimizations on
the command line by defining the ICP_COPTS variable when calling `make`.

By default, optimize for the host platform.

Finally, add a check in the makefile to ensure that required cpu flags
are present and add them if they're missing.