mk: add DPDK machine defines
DPDK header files, specifically rte_memcpy.h, need DPDK's machine cpu
defines to be set in order to enable the most performant version.
Update our DPDK makefiles to set the appropriate defines.

Note: this change only adds a subset of the defines, as that is all that
is necessary for our included headers.