modules/socket: improve ptrace handling
Extract server ptrace handling into a dedicated compilation module,
process_control_linux.cpp, so that both client AND server may use it.

Add proper ptrace handling to the client so that it can explicitly
enable tracing from the server.

Finally, add the ability to check/enable the CAP_SYS_PTRACE capability
if it is needed and log an appropriate message if it cannot be enabled.