tests: grade UDP on a curve
When testing UDP dataplane traffic with `nc`, data can be lost due to
the fact that `nc` uses a 16k buffer for input, attempts to transmit the
buffer in one `send()` call, and the stack does not (currently) generate
fragments.  As a result, itermittent test failues can occur on slower
test systems due to `nc` generating "jumbo" frames inside the stack.

Hence, lower the read/write thresholds for UDP so that any reasonable
transmit/receive performance will be considered a pass.