modules/timesync: add timesync implementation
The timesync module provides a clock for internal services to use for
timestamping events.  By default, the local CPU cycle counter is used
to generate timestamps.  The module can also be synchronized to a
user specified, network time source to provide a mapping between counter
values and wall clock time.

Since timesync depends on digestible, add a digestible makefile so that
we can use it.

Finally, add the timesync module to the openperf target so that clients
can use it.