modules/packetio: Fix pbuf assert() with indirect/external mbufs
The PBUF_POOL type sets PBUF_TYPE_FLAG_STRUCT_DATA_CONTIGUOUS
which indicates that the pbuf header is contiguous to the pbuf
data and enables extra error checking.

This will not be true for external/indirect mbufs which
reference data which is not contiguous to the mbuf so need to
clear PBUF_TYPE_FLAG_STRUCT_DATA_CONTIGUOUS for those cases.