Add packet protocol code
The packet protocol code is divided into two components, which are both
primarily automatcally generated.

The contents of src/lib/packet contain a library for accessing fields
of standard protocols in an endian agnostic manner.  Protocol
definitions are written in YAML with c++ code generated by a code
generator written in python.

The contents of src/packet/protocol contain code to translate packet
library code into swagger objects.  This is done by automatically
generating the swagger specification and translation code from the YAML
defintions in the packet library.

The separation is to allow clients to use the packet library without
having to pull in all of the swagger modules along with it.

Finally, update the top level Makefile so that all automatically
generated code can be updated correctly by executing `make codegen`.