modules/packetio: add headroom to DPDK packet pools
By default, DPDK points the packet buffer data pointer at an offset into
the packet buffer equal to the headroom value.  Add this offset to the
size of packet pool buffers so that we don't end up with data pointers
pointing to the end of the buffer when using small packet sizes.