modules/packetio: add tx checksum offloads
Add a tx_offload function for packetio's packet buffers. This allows
clients to specify the headers and their offsets so that checksum
offloading can be enabled in the backend. Update clients.

Note: this change adds a strongly typed flag implementation to allow
clients to combine multiple enum values into a single "flags" type that
is congruent to the value needed by the backend.