circleci: speed up build
Split the CircleCI workflow into multiple jobs to optimize the build for
wall clock time.

Containers are significantly faster than virtual machines for code
compilation, so do all building in a container. Use the virtual machine
for acceptance tests, so that we can use a privileged containe, and for
static analysis, so that we can use more memory.