circleci: speed up build
Split the CircleCI workflow into multiple jobs to optimize the build for
wall clock time.

Containers are signficantly faster than virtual machines, so do all code
building and analysis in a native container. Only use a virtual machine
to run the acceptance tests, as we need to run them with elevated
privileges.