circleci: speed up build
Split the CircleCI workflow into three independent components:
acceptance tests, unit tests, and static anlysis. Run all three in
parallel for continuous integration builds.

Unfortunately, running the build and acceptance tests inside a container
is still a bottleneck. But we need privileges that are not available in
regular CircleCI containers and the capability bits do not appear to be
supported on any mechanism that moves files between machines and
containers. Hence, I don't know of any good alternatives to building and
running the acceptance tests in a privileged container.  Sigh.