packet/capture: Fix race condition with delete after pcap transfer
The pcap transfer is done from a separate thread from the HTTP
handler.  There was a race condition between the transfer thread
doing cleanup and the HTTP thread processing a delete request.

This was fixed by delaying setting the transfer m_done flag until
after the transfer thread is done and exposing a transfer_active
flag in the capture and result objects which can be checked by
clients to determine when it is okay to delete the capture and/or
result object.