packet/capture: Fix race condition with delete after pcap transfer
The pcap transfer is done from a separate thread from the HTTP
handler.  There was a race condition between the transfer thread
doing cleanup and the HTTP thread processing a delete request.

This was fixed by storing objects which can not be deleted
in a temporary trash storage collection and doing garbage
collection when transfers are completed.