modules/api: don't block startup
When started with a configuration file, the API startup function
attempts to populate the REST API with the contents of the file.
Previously, this was done in the main thread, which caused a
potential ordering issue when the API module was started before a
module it needed to configure. Fix the issue by spawning a separate
configuration thread to do the configuration so that other modules
aren't blocked during startup.

Also, verify that resource paths are ready before attempting to use
them. This avoids a race between the API config thread and other
modules.