packet/analyzer: add reset capability
Add the ability to reset a running analyzer. This replaces the result
object of a running analyzer with a new intance of one and effectively
performs a result counter reset. The REST "reset" method has the same
semantics as the "start" method. A result is returned on success.

Also, add some new AAT's to verify functionality.