modules/packet: make protocol counters shareable
Extract the protocol counters from the analyzer; move both the code and
REST API definitions into a top level module and file, respectively.
We need to separate the protocol counters from the analyzer so that we
may use them in the generator for a forthcoming change.

Additionally, do some cleanup during this transition:
* Rename the layer 4 counters: protocol --> transport
* Rename the flow frame counter: counter --> frame_counter

Update swagger specifictations, analyzer code, and unit tests.