packet/bpf: AAT work around for non-signature payload fill issue
AAT tests were failing in circleci builds because one of the
tests was sending signature and non-signature packets with
the same size.

Currently it doesn't look like the generator zero fills the
payload data.  This causes issues when a mbuf previously
used for signature traffic is recycled for non-signature
traffic because it will still have the stale signature
in it.

In the circleci build the VM only has 2 cores which seems
to cause the mbufs get recycled a lot faster.  In some
cases the same mbuf is used for all packets in the test.

Using different packet sizes for signature and non-signature
packets works aroud the issue because the non-signature packets
are smaller so the signature isn't seen in the payload.
This also works if it was bigger too because the signature
is detected at a fixed offset from the end of the packet.