modules/packetio: add ability to skip packetio init
Update the packetio initialization code to bail out of the
initialization process if the driver is unusable or the user explicitly
disables it. Add a function so that clients can determine whether the
packetio module is enabled or not. Finally, update dependent modules to
check the packetio module state so they can skip their own
initializations, if necessary.

Also, add some AAT's to verify command line argument processing and
module disablement.