tests/aat: add packet tests
Add a test script, packet_ping.py, that performs pings using an
AF_PACKET socket so that we can verify the functionality of AF_PACKET
sockets in openperf. Add new test cases that use the script to the
dataplane tests. Finally, add the `dpkt` package to requirements.txt, as
it is used by the new script to generate packets. The existing `scapy`
package was unusable for this purpose, as it uses the `/proc` filesystem
to determine interface names during its initialization and is
incompatible with `openperf` LD_PRELOAD usage.

Note: the dataplane interface names needed to be changed to be less than
16 characters. This is the name length limit in the `ifreq` struct which
is used by ioctls for various interface related queries.