packet/analyzer: use 0 to indiciate no stream id in flows
Originally, we used the max uint32_t value to indicate the absence of a
stream id in flows.  This was a bad choice, as 1) that could be a valid
stream id and 2) we can never generate a stream id of 0.  Hence, use 0
instead of 0xffffffff when no stream id is present.