dpdk: enable Mellanox and Hyper-V netvsc drivers
Mellanox drivers will only be enabled in the build if the required rdma/infiniband
drivers are installed on the system (libmlx4, libmlx5).

When Mellanox drivers are enabled they are compliled with dlopen stub support
enabled so that rdma/infiniband drivers must only be installed on the
system when using Mellanox DPDK drivers.