modules/socket: fix client channel eventfd leak
client channel objects were not getting deleted because hashtab value
object deletion is not done in delete.  Deletion is only done in
hashtab garbage collection which was not being done.

Added hashtab garbage collection with arbitrarily chosen threshold of
10 items.  eventfds are closed immediately to ensure expected
socket behavior of immediately closing the fd, but deletion is batched
to reduce exclusive locking.