modules/packetio: fix off by 1 error
Oops. DPDK packet pools are constrained to a Mersenne number in size. If
a client requests a power of two size, then the actual size might be 1
less instead of doubling the pool size. As a result, we need to use the
actual size of the resulting pool instead of the requested size when
looking for compatible memory pools to share.