modules/socket: fix socket close() with  LD_PRELOAD
Previous fix for leaking channel objects caused isses when used
with LD_PRELOAD because overridden close() was trying to grab
the non-recursive mutex again.

This change moves eventfd close() outside of mutex so the mutex
will be available.

The ided_channel objects were also wrapped with a shared_ptr<>
to prevent channel deletion while the channel object is still
in use.