modules/socket: fix threaded client usage
Since openperf clients may be threaded, the client code needs to handle
calls in a thread safe manner. This changelist fixes threaded client
usage in two ways:

1) Use a single global instance for the client hooks. Thread local
   instances can't actually construct themselves due to a catch 22
   situation with libc functions.

2) Add a lock (eww) to the RPC socket so that client theads can safely
   share it.

Additionally, change the channel hashtab to a client member instead of a
singleton. There's no reason for it to exist outside the client now.