modules/socket: make socket stats REST API consistent
This change makes the socket stats REST API consistent with
other REST APIs.  The all socket stats URL is /sockets and
a single socket's stats may be retrieved using the URL
/sockets/{id}.  Where "id" is the unique socket ID used to
indentify the socket.

The socket ID is derived from the PCB address which is
exposed in the socket stats as a UUID string.

Support was also added for retrieving socket stats from PCBs which
are known by sockets but not in lwip TCP/UDP PCB lists.

This includes unregistred TCP/UDP sockets and other types of sockets
(ip, raw and packet).