lib/regurgitate: change scaling algorithm
Replace the "k2" cluster scaling algorithm from the original t-digest
paper with the Folly algorithm. This algorithm is quadratic instead of
logarithmic, so the cluster values are computationally easier to
compute. It also generates less tight bounds near the extremes, so the
resulting digest is easier to map to a traditional histogram.

Additionally, drop a couple of problematic unit tests that were
sensitive to the cluster algorithm due to float resolution issues.

Finally, loosen the AAT check to avoid float resolution issues.