src: fix x86 instruction set detection
The `xgetbv` instruction necessary for determining AVX support isn't
available on all processors. Use the OSXSAVE bit as a proxy to determine
if the host CPU supports it or not before attempting to execute it.

Update SSE2/SSE4 detection to not require OSXSAVE, since it could be
unset on platforms with support SSE instructions.