modules/timesync: fix overflow in error calculation
The clock attempts to use the best NTP exchanges to measure timer
frequency. The level shift detection mechanism allows the clock to
adjust the definition of "best" due to network issues over time.
However, that mechanism's implementation had a bug. Timestamps below the
new, minimum round trip time from level shifting need to be removed from
both the round trip time digest *AND* the timestamp history. Failure to
prune the values from the history caused the frequency measurement
algorithms to use timestamps with RTT values below the level set minimum.
This caused an integer overflow when calcuating the error, which resulted
in a large, and incorrect, error value.