packet/stack: fix internal config pointers
The LwIP netif structure only has a pointer for the interface hostname.
As a result, we need to use a pointer to the actual data and not a
temporary reference when "setting" the hostname.

This change allows the stack to advertise the hostname set by the user
when requesting DHCP addresses.