mk: fix compiler optimizations
Since we now have explicit OP_EXTRA_* flags for compiler/linker options
make the OP_COPTS options strictly an internal one. This fixes the issue
where optimizations were not being applied to builds when OP_COPTS
needed to be modified, as the `override` tag wasn't used everywhere it
need to be.