packet/generator: base tx duration on expected tx rate
When using an explicit transmit duration, the generator calculates an
expected transmit frame count and uses that value to determine when to
stop transmitting packets. However, if the generator was either CPU or
network bound, then the intended transmit rate couldn't actually be hit
and the generator would transmit for longer that it's configured
duration.

Fix this problem by using the expected transmit count instead of the
actual transmit count when determining when to stop the generator when
durations are used.

Add an AAT to verify. Resolves #515