tests/aat: add force-unlink option to drop test
The generator drop test runs after an AAT suite that sometimes doesn't
properly clean up after itself. As a result, add the `force-unlink`
option to the generator drop configuration so that it will always run.

The issue in the previous suite is related to a mostly harmless shutdown
race that should be fixed, but hasn't been... yet.