api: modify cpu schema
Add a new object to the CPU schema, CpuGeneratorCoreLoad. This object
describes both the instruction set and data type used for CPU load
generation. Use this object for both configuration and results.
Previously, the result objects did not provide any indication as to what
type of CPU load was being generated. This change removes that
ambiguity.

Additionally, this change updates generated code.