tests/unit: fix capture buffer tests when clock update is slow
The unit test was using the system clock to generate a starting
timestamp in the fill_capture_buffer_ipv4() function.

The test can fail if the system clock does not increment fast
enough so that the fill_capture_buffer_ipv4() generates
incrementing timestamps.

Test failures were observed when running on WSL2/Hyper-V.