modules/packetio: DPDK lcore cpuset support
This change fixes issues when lcores use more than 1 CPU (cpuset).

When an lcore cpuset spans multiple NUMA nodes the lcore is not
assigned to a socket ID.  The stack code was using the invalid
socket ID from the lcore thread to determine the mempool and
was causing a crash.

The stack code was changed to use socket ID 0 when the socket ID is
not valid.  Additional code was added to handle cases where the
device NUMA affiliation does not match the CPU NUMA node.

The packetio driver init was also made lcore cpuset aware so it
removes cpus from the default cpuset based off of the cpuset assigned
to the lcore instead of assuming the lcore ID is the CPU ID.