packet/stack: change IPv6 gateway route precedence
Change the precendence of the configured IPv6 gateway address so it has
higher precedence than routers found through discovery.

This provides more flexibility because not configuring the
gateway address provides the original behavior so functionality
can be turned on/off.  The other way it depends if there is
a IPv6 router on the network or not.

lwip also currently does not check the prefixes the router is advertising so
it is safer to use a configured default route if one is provided.

lwip did not have IPv6 gateway support to begin with, so this change
is fixing code I previously added...