modules/packetio: add ports-required option
add ports-required option which causes openperf to exit if it
does not find all ports specified in the configuration.

This option was added to support recovery when running in container
environments (kubernetes).  There were cases where the DPDK library
fails to intialize a VF interface due to the interface being in a
reset state and cases where the PF driver does not respond to admin
messages.  This appears to be a transient condition when the system
is under load and the VF interface eventually seems to recover.

Previously without this option enabled, openperf will log a warning
and startup without the intefaces.  This makes it hard to detect
issues and recover from it.

With the ports-required option enabled, openperf will exit if it
does not find all of the ports it expects which makes it easier to
detect a problem and try again.