modules/block: add suport for faster stopping
This change makes the block IO generator break out of
a worker spin loop which is taking a long time when
the block generator is stopping.

There were some cases where stopping the block generator was taking
several minutes on network storage with large block sizes and
larger queue depths.