modules/memory: fix read-only test without mlock (#575)
* modules/memory: fix read-only test without mlock

This change fixes the memory generator read-only mode when mlock()
is not supported/allowed.

When mlock() is not supported/allowed, a read-only test does not appear
to be accessing RAM because the read rates are much higher than the
write rates.  This is likely because memory is not being allocated for the mmap.

This change fixes the issue by touching the mmap memory, which forces
it to get allocated.

* update circleci image

The previous image is no longer available.